Added BaseCompiled and BaseInterpreted execution strategies https://github.com/SoftUni-Internal/interactive-issues/issues/250
