--- OJS.Workers.Common/Exceptions/CompilationFailedException.cs
@@ -1,12 +1,0 @@
-﻿namespace OJS.Workers.Common.Exceptions
-{
-    using System;
-
-    public class CompilationFailedException : Exception
-    {
-        public CompilationFailedException(string compilerComment)
-            : base(compilerComment)
-        {
-        }
-    }
-}
