+++ OJS.Workers.ExecutionStrategies/BaseCodeExecutionStrategy.cs
@@ -1,17 +1,14 @@
 ﻿namespace OJS.Workers.ExecutionStrategies
 {
     using System;
-    using System.IO;
 
     using OJS.Workers.Common;
     using OJS.Workers.Common.Extensions;
-    using OJS.Workers.Common.Helpers;
     using OJS.Workers.Common.Models;
-    using OJS.Workers.Compilers;
     using OJS.Workers.ExecutionStrategies.Models;
     using OJS.Workers.Executors;
 
-    public abstract class BaseCodeExecutionStrategy : BaseExecutionStrategy
+    public class BaseCodeExecutionStrategy : BaseExecutionStrategy
     {
         protected const string RemoveMacFolderPattern = "__MACOSX/*";
 
@@ -34,55 +31,6 @@
         protected IExecutor CreateExecutor(ProcessExecutorType processExecutorType)
             => this.ProcessExecutorFactory
                 .CreateProcessExecutor(this.BaseTimeUsed, this.BaseMemoryUsed, processExecutorType);
-
-        protected IExecutionResult<TestResult> CompileExecuteAndCheck(
-            IExecutionContext<TestsInputModel> executionContext,
-            IExecutionResult<TestResult> result,
-            Func<CompilerType, string> getCompilerPathFunc,
-            IExecutor executor,
-            bool useSystemEncoding = true,
-            bool dependOnExitCodeForRunTimeError = false)
-        {
-            // Compile the file
-            var compileResult = this.ExecuteCompiling(
-                executionContext,
-                getCompilerPathFunc,
-                result);
-
-            if (!compileResult.IsCompiledSuccessfully)
-            {
-                return result.CompilationFail(compileResult.CompilerComment);
-            }
-
-            var outputFile = compileResult.OutputFile;
-
-            // Execute and check each test
-            var checker = executionContext.Input.GetChecker();
-
-            foreach (var test in executionContext.Input.Tests)
-            {
-                var processExecutionResult = executor.Execute(
-                    outputFile,
-                    test.Input,
-                    executionContext.TimeLimit,
-                    executionContext.MemoryLimit,
-                    null,
-                    null,
-                    false,
-                    useSystemEncoding,
-                    dependOnExitCodeForRunTimeError);
-
-                var testResult = this.CheckAndGetTestResult(
-                    test,
-                    processExecutionResult,
-                    checker,
-                    processExecutionResult.ReceivedOutput);
-
-                result.Results.Add(testResult);
-            }
-
-            return result;
-        }
 
         protected TestResult CheckAndGetTestResult(
             TestContext test,
@@ -136,48 +84,5 @@
                     ? processExecutionResult.ReceivedOutput
                     : processExecutionResult.ErrorOutput
             };
-
-        protected CompileResult ExecuteCompiling<TInput, TResult>(
-            IExecutionContext<TInput> executionContext,
-            Func<CompilerType, string> getCompilerPathFunc,
-            IExecutionResult<TResult> result)
-            where TResult : ISingleCodeRunResult, new()
-        {
-            var submissionFilePath = string.IsNullOrEmpty(executionContext.AllowedFileExtensions)
-                ? FileHelpers.SaveStringToTempFile(this.WorkingDirectory, executionContext.Code)
-                : FileHelpers.SaveByteArrayToTempFile(this.WorkingDirectory, executionContext.FileContent);
-
-            var compilerPath = getCompilerPathFunc(executionContext.CompilerType);
-
-            var compileResult = this.Compile(
-                executionContext.CompilerType,
-                compilerPath,
-                executionContext.AdditionalCompilerArguments,
-                submissionFilePath);
-
-            return compileResult;
-        }
-
-        protected virtual CompileResult Compile(
-            CompilerType compilerType,
-            string compilerPath,
-            string compilerArguments,
-            string submissionFilePath)
-        {
-            if (compilerType == CompilerType.None)
-            {
-                return new CompileResult(true, null) { OutputFile = submissionFilePath };
-            }
-
-            if (!File.Exists(compilerPath))
-            {
-                throw new ArgumentException($"Compiler not found in: {compilerPath}", nameof(compilerPath));
-            }
-
-            var compiler = Compiler.CreateCompiler(compilerType);
-            var compilerResult = compiler.Compile(compilerPath, submissionFilePath, compilerArguments);
-
-            return compilerResult;
-        }
     }
 }
