+++ OJS.Workers.ExecutionStrategies/CSharp/DotNetCoreUnitTestsExecutionStrategy.cs
@@ -6,7 +6,6 @@
     using System.Linq;
 
     using OJS.Workers.Common;
-    using OJS.Workers.Common.Extensions;
     using OJS.Workers.Common.Helpers;
     using OJS.Workers.Common.Models;
     using OJS.Workers.ExecutionStrategies.Extensions;
@@ -103,9 +102,12 @@
                     executionContext.AdditionalCompilerArguments,
                     consoleRunnerPath);
 
+                result.IsCompiledSuccessfully = compilerResult.IsCompiledSuccessfully;
+                result.CompilerComment = compilerResult.CompilerComment;
+
                 if (!compilerResult.IsCompiledSuccessfully)
                 {
-                    return result.CompilationFail(compilerResult.CompilerComment);
+                    return result;
                 }
 
                 // Delete tests before execution so the user can't acces them
