+++ OJS.Workers.ExecutionStrategies/PHP/PhpCgiExecuteAndCheckExecutionStrategy.cs
@@ -8,7 +8,7 @@
     using OJS.Workers.ExecutionStrategies.Models;
     using OJS.Workers.Executors;
 
-    public class PhpCgiExecuteAndCheckExecutionStrategy : BaseCodeExecutionStrategy
+    public class PhpCgiExecuteAndCheckExecutionStrategy : BaseInterpretedCodeExecutionStrategy
     {
         private const string FileToExecuteOption = "--file";
 
@@ -33,9 +33,6 @@
             IExecutionContext<TestsInputModel> executionContext,
             IExecutionResult<TestResult> result)
         {
-            // PHP code is not compiled
-            result.IsCompiledSuccessfully = true;
-
             var codeSavePath = FileHelpers.SaveStringToTempFile(this.WorkingDirectory, executionContext.Code);
 
             // Process the submission and check each test
